F-18 split _tuning/engine.py into leaf modules
engine.py (1,830 lines) is now a ~700-line orchestrator: params,
splitters, metrics, grid_random, refit leaf modules plus a
strategies/ package (halving, optuna lazy loader, shared runner).
Pure code movement - public TuningCalculator/TuningApplier surface
unchanged, test-pinned methods kept as one-line delegates, stale
test pins retargeted to the leaf modules.
