+++ requirements-ci.txt
@@ -31,7 +31,7 @@ scipy>=1.13.1
 imbalanced-learn>=0.14.1,<1.0.0
 joblib>=1.3.0
 optuna>=3.0.0                        # Imported by skyulf-core/tests/test_tuning_engine.py — needed for ty type-resolution in CI
-optuna-integration>=3.0.0            # HAS_OPTUNA also requires this fallback import path to succeed (see engine.py)
+optuna-integration>=3.0.0            # HAS_OPTUNA also requires this fallback import path to succeed (see skyulf-core/skyulf/modeling/_tuning/strategies/optuna.py)
 cmaes>=0.10.0                        # Required by optuna's CmaEsSampler, exercised in test_tuning_engine.py
 matplotlib>=3.7.0                    # Imported by skyulf-core/tests/test_profiling_visualizer.py — needed for ty type-resolution in CI
 rich>=13.0.0                         # Imported by skyulf-core/tests/test_profiling_visualizer.py — needed for ty type-resolution in CI
